requirements.txt
# requirements.txt
# Author: Rushy Panchal
# Date: March 8th, 2017
# Description: Default Python requirements file.

Django==1.10
dj-database-url==0.4.2
django-cas-ng==3.5.4
djangorestframework==3.6.2
psycopg2==2.6.1
python-cas==1.1.0
PyYAML==3.11
six==1.10.0
